refactor: [SCRUM-136] Consolidate auth API calls into single index.js and Update import paths in LogoutModal
