+++ src/pages/LogoutModal.jsx
@@ -2,7 +2,7 @@ import { Modal, Button, Typography, message } from 'antd'
 import { useNavigate } from 'react-router-dom'
 import { useDispatch } from 'react-redux'
 import { logout } from '@app/providers/reducer/auth/authSlice'
-import { useLogout } from '@features/auth/authen-user/api'
+import { useLogout } from '@features/auth/api'
 import { useState } from 'react'
 import { ACCESS_TOKEN } from '@shared/lib/constants/auth'
 import { jwtDecode } from 'jwt-decode'
